fix(ci): resolve compilation errors in Phase 5 notifications
Fixed 4 compilation errors:
- CS0266: Cast decimal division result to double in BudgetAlertService
- CS0117: Replace Cron.DailyAt with standard cron expression "0 8 * * *"
- CS0117: Replace Color.TextSecondary with Color.Default in NotificationCenter
- CS0117: Replace Color.TextSecondary with Color.Default in NotificationPreferences

These changes address the remaining CI build failures from the Phase 5 notifications implementation.

+++ src/MyDesk.Web/Services/BudgetAlertService.cs
@@ -69,7 +69,7 @@ public class BudgetAlertService
             int budgetId = (int)budget["BudgetId"];
 
             decimal totalCommitted = spent + encumbered;
-            double usagePercentage = allocated > 0 ? (totalCommitted / allocated) * 100 : 0;
+            double usagePercentage = allocated > 0 ? (double)((totalCommitted / allocated) * 100) : 0;
 
             _logger?.LogInformation(
                 "Budget usage: {Usage}% ({Spent} of {Allocated})",
